Add submit event on successful completion
As far as I can tell, requires detecting the congratulatory modal. We're
being generous here and only registering a submit for a successful solve
- adding submissions for unsuccessful solves is still to-do. (A little
harder because if you already have a full but incorrect grid and you
change a square and are still incorrect, there's no new modal.) The
result is that submit events are always success: true for now.