Refactor - depend on, but don't yet use, a storage key
Eventually we want to store the record of a solve with chrome.storage.
This refactors the code a little bit so that there is a single `record`
object, and so that observers are only attached as part of a callback
after chrome.storage is checked. However, we do not yet actually use the
stored value or write to it at any point.