Add handling for navigation away/to page (partly done)
When navigating away from a puzzle, add a stop event if necessary and
store the record on the spot. When completing a puzzle successfully,
store the record. Add a little check to make sure that we don't
re-record fill we made on a previous solve, not the most robust but it
seems to work.

Code's kind of messy but I'll reorganize it later.