Move store management to the service worker
This is a little cleaner, as I would expect this code to be reusable
across any other websites we'd support in the future besides the NYT. It
also minimizes the amount of work being done in the content script, just
fire off the message and go on with life. And most importantly, it seems
to fix an issue I'd seen where the onbeforeunload handler didn't work
correctly, got complaints about "Execution context invalidated". It
seems to be OK if all I need to do is send a message.