Add basic options screen
The options screen supports a solver name, which is now also logged as
part of the record, and setting the default frequency with which we
flush events.

I removed the ability to enable/disable trackers. I don't think it makes
sense to support this, if you don't want tracking then you can just
disable the extension itself.