Service worker cleanup
Clean up the code a little bit in the syntax, so that it's clear that
there are two things - one to set default preferences on first install,
and the "main" bit that listens to messages and takes action.

Remove deleteRecordByKey, I don't think it justifies having a function
of its own.