NYT refactor - single observer
Major refactor where instead of creating several MutationObservers,
including one for each cell, we just create a single observer and have
a more complex single callback that determines what events to record.
The callback can record multiple events, or none at all.

The reason for this is that with separate observers, we didn't have
strong guarantees about the order in which different callbacks would
fire if they were fired at the same time. The clearest example of this
is that since the NYT will auto-submit a correct puzzle on final fill,
sometimes the "submit"-generating callback would fire before the
callback that filled the final square.

Also to help address the above issue, I've defined an order in which
concurrent events should be added.

I've pared back the number of "magic word" classes that we rely on, and
it should be consistent with the NYT-specific documentation now.

I've hopefully offset this complexity by making the code cleaner and
simpler to read.

Somewhat relatedly, I've temporarily scrapped the "highlight" event. I
think we want this to be keyed to clues rather than squares. The number
of extra events generated by tracking every highlighted square is quite
large and not that informative; it's really "what clue is the solver
focused on" that we want to keep track of.