Log clue selection
Add an event "selectClue" that indicates that a certain clue is being
looked at closely by the solver.

This doesn't add too much to the logs, so simplify the log level
settings to just two choices: basic (no navigation) and full (with
navigation), and default to full.

Displaces the old approach of highlighting squares.