Start adding puzzle SVG generation
The eventual plan is to make it a file to download, but for now it just
inserts the SVG directly into the options page.

Right now initial state and basic solve is done, but there are several
todos: clue highlighting, progress (the "complete" text and the timer),
and de-hardcoding the image parameters.