Base NYT storageKey off of title and date
URL is not a good identifier unfortunately, there are URLs that are
"today's puzzle" for the mini that always reflect what the latest puzzle
is. I would not be surprised if other sites have this as well, so URLs
are not the right way to define a storage key.

The storage key is now based off of the puzzle title and date. I've
rearchitected the code a bit so that we don't kind of globally rely on
the definition of storageKey, instead passing it around as an argument.