Use SHA1 to define storage keys
Base the SHA1 digest on a puzzle's name, date, and byline (as in
previous commits, I'm moving away from URL as it is generally possible
for it to correspond to different puzzles every day).

I don't think this is necessary for Chrome storage, but it's helpful for
the options page where we tag record rows with `id` equal to the storage
key. And this keeps the storage keys a little less unwieldy anyway.