Fix event ordering
I ran into situations where events were not necessarily appearing in the
right order - a successful submit event concurrent with another event
was sometimes not the last event to be recorded. I suspect that we may
get multiple callback firings even with events that are deemed to occur
at the same time (and I'm not sure in what order those would be fired).

Thus, I revamped the addition of events to actually insert the event in
the correct order in the record's event history.