Refactor into single appSettingsInfo const
Move validation into appSettings, and update appSettings so that it is
an array (in a well-defined order); this will be helpful when I refactor
the options page to generate input elements based on this. Likewise,
within each element of appSettings, the settings key now contains an
array of information about each setting.

Finally, rename appSettings => appSettingsInfo, a little less misleading
- this contains info about what settings are available, not the settings
themselves.