Derive puzzle storageKey from structure; general code cleanup
Implement a different approach where we derive a puzzle identifier from
the structure of the puzzle itself.

Puzzle names alone are not enough. Dates can be in a variety of formats,
and in the future we may want to attempt parsing the dates; having the
storage key derived from the date field means that a parsed date
registers as a different puzzle. Bylines can change if the author or
editor's name changes. Thus, the old approach of deriving a puzzle
identifier from these three fields is not good enough.

This approach derives the puzzle identifier by digesting the structure
of the puzzle itself: the fillable squares and the clue text. Arguably
if any one of these changes, you really do have a different puzzle.

Also do some general code cleanup.