Switch to SHA1 for puzzle identifiers
Use SHA1. Compute it with crypto.subtle, and get rid of hashCode. This
requires us to place the "main" code execution in the NYT content script
into a separate async function, which honestly is not a bad idea
architecturally anyway (at least the separate function part).

Update documentation to reflect this.